feat: add `merge_base::octopus()`.
